style: move config outside the app folder
