--- app/configs/__init__.py
@@ -1,1 +1,0 @@
-from .develop import *
